+++ vm/src/vm/runners/cairo_runner.rs
@@ -80,6 +80,7 @@ pub enum CairoArg {
     Composed(Vec<CairoArg>),
 }
 
+
 // Converts a vector of values into an array-style Cairo argument.
 impl<T> From<Vec<T>> for CairoArg
 where
@@ -1679,6 +1680,7 @@ impl CairoRunner {
             })
             .collect()
     }
+
 }
 
 //* ----------------------